First draft on full pipeline
A couple of things….

1. I decided just to throw all of the packages in the git repo. This
may not be the most elegant way of doing it but I think it’s better
than what we had. Let me know if you guys don’t like this.

2. I haven’t put the QC checks in there yet.

3. VDJFasta writes to the same folder as the .fasta files are put. Not
sure how to fix that.

4. I just rewrote all the code and put it into one executable. Party
that was for me, so I could see how it all went together. This may not
be the correct way to do it. If it’s better to have many sub functions
let me know, and I’ll fix it. Robby, I’m not sure if this is important
for your flagging.

5. If any of this is not done right, just tell me and I’ll fix it. I
need to learn the best practices way of organizing this.

6. I’ll do documentation after everyone agrees this is ok.