Added ete2 obj to rep seq
Added two new “features” in the self.features dict for the rep_seq
object. The first, called “tree_dict”, is a dictionary where each key
is a VJ pair and each associated object is an ete2 tree object with the
output from Immunitree. The second one, called “pruned_tree_dict”, is
exactly the same except the leaves are pruned off. Also added a
function in rep_stats that converts the node.txt output of Immunitree
to ete2 format.